Create BC_Training.py
A program to train (and validate) four models (Random Forest, Support Vector Machines, K-Nearest Neighbour, and Multi-Layer Perceprton Neural Networks) with the selected features. A random sample is selected for trainign (70% of the data base) and the remaining data (30%) is left for validation. Each sample is used in the four models, before generating a new one (this is done 10 times). All results are stored in csv files, showing the sample used for validation, the output and the expected output.